refactor : LadderView -> LadderResult로 변경, domain으로 위치 이동
