+++ src/main/java/ladder/controller/Controller.java
@@ -1,12 +1,11 @@
 package ladder.controller;
 
 import ladder.domain.game.Ladder;
-import ladder.domain.game.Name;
 import ladder.domain.game.Names;
 import ladder.strategy.ConnectionStrategy;
 import ladder.strategy.RandomConnectionStrategy;
 import ladder.view.InputView;
-import ladder.view.LadderView;
+import ladder.domain.game.LadderResult;
 import ladder.view.ResultView;
 
 import java.util.Scanner;
@@ -30,13 +29,13 @@ public class Controller {
 
         ResultView.printLadder(participants, ladder, goals);
 
-        LadderView ladderView = LadderView.of(participants, ladder.moveAll(goals));
+        LadderResult ladderResult = LadderResult.of(participants, ladder.moveAll(goals));
 
         String inputName;
 
         do {
             inputName = InputView.askResultPerson();
-            ResultView.printGoals(ladderView, inputName);
+            ResultView.printGoals(ladderResult, inputName);
         } while (!inputName.equals(ResultView.RESERVED_WORD_ALL));
 
     }
