+++ src/main/java/ladder/view/ResultView.java
@@ -1,9 +1,6 @@
 package ladder.view;
 
-import ladder.domain.game.Ladder;
-import ladder.domain.game.Line;
-import ladder.domain.game.Name;
-import ladder.domain.game.Names;
+import ladder.domain.game.*;
 
 import java.util.List;
 import java.util.Map;
@@ -33,14 +30,14 @@ public class ResultView {
         showNames(goals.getNames());
     }
 
-    public static void printGoals(LadderView ladderView, String inputName) {
+    public static void printGoals(LadderResult ladderResult, String inputName) {
         System.out.println(System.lineSeparator() + RESULT_GOALS);
 
         if (inputName.equals(RESERVED_WORD_ALL)) {
-            showAllResults(ladderView.getResultOfAll());
+            showAllResults(ladderResult.getResultOfAll());
         }
         if (!inputName.equals(RESERVED_WORD_ALL)) {
-            System.out.println(ladderView.getResultOfOneParticipant(Name.from(inputName)));
+            System.out.println(ladderResult.getResultOfOneParticipant(Name.from(inputName)));
         }
     }
