+++ src/test/java/ladder/view/LadderResultTest.java
@@ -1,12 +1,7 @@
 package ladder.view;
 
-import ladder.domain.game.Ladder;
-import ladder.domain.game.Line;
-import ladder.domain.game.Name;
-import ladder.domain.game.Names;
-import org.assertj.core.api.Assertions;
+import ladder.domain.game.*;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -15,12 +10,11 @@ import java.util.Arrays;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Created By mand2 on 2020-12-09.
  */
-class LadderViewTest {
+class LadderResultTest {
 
     Ladder ladder;
     Names participants;
@@ -58,18 +52,18 @@ class LadderViewTest {
     @ParameterizedTest
     @MethodSource("providedResultOfLadder")
     void 결과값_가져오기_각자(Name inputParticipant, Name expectedGoal) {
-        LadderView ladderView = LadderView.of(participants, ladder.moveAll(goals));
+        LadderResult ladderResult = LadderResult.of(participants, ladder.moveAll(goals));
 
-        assertThat(ladderView.getResultOfOneParticipant(inputParticipant))
+        assertThat(ladderResult.getResultOfOneParticipant(inputParticipant))
                 .isEqualTo(expectedGoal);
     }
 
     @ParameterizedTest
     @MethodSource("providedResultOfLadder")
     void 결과값_가져오기_전체(Name inputParticipant, Name expectedGoal) {
-        LadderView ladderView = LadderView.of(participants, ladder.moveAll(goals));
+        LadderResult ladderResult = LadderResult.of(participants, ladder.moveAll(goals));
 
-        assertThat(ladderView.getResultOfAll().get(inputParticipant))
+        assertThat(ladderResult.getResultOfAll().get(inputParticipant))
                 .isEqualTo(expectedGoal);
     }
